LayoutTests/media/media-source/media-source-real-fastseek-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EVENT(updateend)
RUN(source.duration = 8)
RUN(source.endOfStream())
RUN(video.currentTime = 1)
EVENT(seeked)
EXPECTED (video.currentTime.toFixed(1) == '1') OK
RUN(video.fastSeek(5))
EVENT(seeked)
EXPECTED (video.currentTime.toFixed(1) == '4') OK
RUN(video.currentTime = 3)
EVENT(seeked)
EXPECTED (video.currentTime.toFixed(1) == '3') OK
RUN(video.fastSeek(2))
EVENT(seeked)
EXPECTED (video.currentTime.toFixed(1) == '0') OK
END OF TEST
